locations controller

+++ src/main/java/org/example/dto/LocationPageDTO.java
@@ -5,5 +5,5 @@ import java.util.List;
 public record LocationPageDTO (
         List<LocationWeatherDTO> locationWeatherDTOList,
         int currentPage,
-        Long overallPages
+        int overallPages
 ){}
